Add "batched" and "single" line drawing modes
Turns out a major timesink in approaches lies in calling getImageData
and putImageData. If we perform that for each iteration of drawing
a line, what we end up with is a difference between 200 ms and 3400 ms,
to draw only five hundred lines.

This makes me wonder if it makes sense to treat lines, by the time that
it arrives at the drawLines function, as just one giant array of pixel
values, rather than two loops of arrays of pixel values. Or if there
are any benefits to chunking points as pairs of values rather than
objects with x and y fields.