Separate example from library code
This change separates index.js into two files: one that has just the
drawing functions and one that has the example drawing.

It gets rid of the individual drawLine function in favor of one that
draws an array of lines. The overhead of calling getImageData and
putImageData between lines is just too high.

Small change to the example as well, it now draws against the whole
viewport.